Refactoring and cleanup of service
